+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0-RELEASE (2021-11-30)
 
 1. Add message stability
+2. Add new coordinape round for guest/L1/L2
 
 ## 1.1.1-RELEASE (2021-11-27)